Mesh Generator basis completed and basic movement for testing
I have finished off the basics of the Mesh Generator, it can generate a quad-grid of any size up to 255 length and width. It can also regenerate it's mesh in-scene using Serialized fields, and you can tweak the size of it's Quads.

I have also setup a quick, non-functional movement script for my camera. It's just for debugging purposes and is not really useful towards Perlin generation.