Moved Code to individual members
Rewrote the code for cohesion within the flock members rather than the manager, to allow for more control of individuality. Also added some code for changing the colour to show the state of the member - Currently it only changes once, being blue when spawned but turning red once it starts following another member. I hope to add a system where as they join a group, they will assimilate their colour - so it will be easy to distinguish the various groups that naturally form.