Better Noise, Tweaks and Tinkering
Further improvements of the Noise algorithmn, uses an exponential Smooth function that transforms the terrain to contain both jagged mountainous enviroments and smooth plains in lower areas. Did some testing with the noise function and can tell it is robust and scalable.

Also reconfigured the Noise mesh to use a Material that respects the colour of the vertices, alongside adding (non-functional) UV's to the vertex material. Now it's time for multi-threading and optimisation work.