Tweaks and Smoothing
Added SmoothStep function to the Noise (both Compute Shader and Static class versions), which makes the noise look less cubic. Also did tweaking in regards to the current noise setup, overall not many changes.

I did try and make a function to snap the camera to a chunks vertex if it is below the chunk, but the current implementation is janky and bugged with rapid snapping shifts. The functions are there but are disabled.