Further look into Quadmesh chunk LOD bordering and Skirts
Further work on getting the chunk meshes to line up at LOD seams and each other, got the chunks to generate accurate normals, at least until it hits the chunk border. I could just regenerate the mesh normals using Unities mesh class, but I need to account for issues in the bordering of LOD chunks, plus there were already present bugs with Unity not knowing that the chunk was continuous.

Current Quadtree chunks just shove a skirt of Vertices below while keeping chunks coherent to each other on the Quadtree grid, the skirts are meant to plug holes in LOD borders as they are simple patches in preventing world holes from showing up - this is of course a hack solution, as scrutinising the details of far away chunks will reveal cliff faces that do not look "natural", although for all intents and purposes this is barely an inconvenience.