Box Collider Push for Aidan
Aidan needed Colliders for his boids so here is a quick push for the Terrain featuring the box collider.

NOTE: This is in the Quadtree Test folder of the Procedural Terrain, not the main Procedural Terrain scene, as I have not officially pushed the arguably much better feature rich Quadtree into the main out of laziness.