(Buggy) Colourmap Established
Figured out that I was incorrectly applying the coordinates to the UV specification, as I was parsing a uint for calculation rather than a float. This turned out to be the main gatekeeper to using colours, as it fixed almost all issues present with coding it. Now all that's left is preventing UV seams to show.

I can now use colour on the material, which is the main thing.