Basemap improvements and Steepness Colour
Improvements to the colouring function for the mesh vertices, offloaded colouring to a custom function that inputs a couple of parameters such as the Left and Right colours for interpolation. I also figured out how to account for the steepness of the mesh, by parsing in the regenerated normals into the buffer I can figure out within the CS the normal/steepness of the terrain and assign a third "steepness colour" which is then interpolated with the base.

This allows me to specify a colourmap with terrain that can interpolate through both height and steepness of the terrain, though I could further improve this by adding in the hash table and integrating noise offset to the colours to add more variation and randomness.