Further colormap work and basic Instantiation implementation
Beginning work on Object Instancing to allow features such as grass and trees to be generated depending on the terrain, started a test scene where I'll try and generate a bunch of grass objects around the player using Compute Shaders.

Optimally this will be a dynamic loading system while retaining moderate performance and allow interesting features to spawn based on Terrain parameters such as height and steepness.