Prefab Instancing and very basic map objects
Doing some cursory tests on basic map objects. Unity enables you to save GameObjects as a "prefab", which can then be used to instantiate GameObjects in C# with duplicate properties (or modify the resulting properties).

I made a "tree" prefab which is basically just a big box, which contains collider and scale data. Additionally, made a basic system of generating roughly a thousand objects in a 500 meter radius around the player, which seems rather performant.

My next step is to improve the system with overhead, and then figure out how to pair it with the terrain generation. Then it's a matter of maybe looking into Corroutines to spread out object generation over frames, and then looking into a better system overall.