Object Generation functional, Settings Serialization
I have successfully integrated a system that automatically generates and places down prefabs of "trees" within a three LOD distance of the player, including draw-checks based on the objects distance to the player and parent-object linking.

This means trees now generate across terrain, and due to the use of Perlin Nosie in their generation algorithm, the tree's are not completely random - they are deterministic based on the map seed, and can regenerate if the chunk they originate from is deleted.

I have also serialized all the primary settings of my world generator, allowing quick tweaks and changes in regards to Quadtree parameters, chunk mesh parameters, and prefab parameters, alongside setting a seed.

All in all, this is considerable progress and I believe this is good enough for Mikey and Aidan to utilise in their Boid and Navmesh.